requirements.txt
requests
python-dotenv
yfinance
ta